+++ src/hardshell/checks/base.py
@@ -1,24 +1,30 @@
 from dataclasses import dataclass, field
 from typing import List, Optional
+from src.hardshell.common.logging import logger
 
 
+@dataclass
 class BaseCheck:
-    def __init__(
-        self,
-        category: Optional[str],
-        check_id: Optional[str],
-        check_name: Optional[str],
-        # check_results: List[str],
-        check_subtype: Optional[str],
-        check_type: Optional[str],
-        depends_on: Optional[List[str]],
-        valid_os: Optional[List[str]],
-    ):
-        self.category = category
-        self.check_id = check_id
-        self.check_name = check_name
-        # self.check_results = check_results
-        self.check_subtype = check_subtype
-        self.check_type = check_type
-        self.depends_on = depends_on
-        self.valid_os = valid_os
+    category: Optional[str] = None
+    check_id: Optional[str] = None
+    check_name: Optional[str] = None
+    check_results: List[dict] = field(default_factory=list)
+    check_pass: List[str] = field(default_factory=list)
+    check_fail: List[str] = field(default_factory=list)
+    check_subtype: Optional[str] = None
+    check_type: Optional[str] = None
+    depends_on: List[str] = field(default_factory=list)
+    valid_os: List[str] = field(default_factory=list)
+
+    def set_result(
+        self, check_id, check_name, check_result, check_subtype, check_type
+    ) -> None:
+        self.check_results.append(
+            {
+                "id": check_id,
+                "name": check_name,
+                "result": check_result,
+                "subtype": check_subtype,
+                "type": check_type,
+            }
+        )