+++ src/hardshell/common/common.py
@@ -5,8 +5,8 @@ import platform
 import re
 from pathlib import Path
 from typing import Callable, Dict, Union
-
 from src.hardshell import __name__, __version__
+from src.hardshell.common.logging import logger
 
 
 def detect_admin() -> bool:
@@ -143,8 +143,8 @@ def find_pattern_in_file(path, pattern):
     return pattern_found, pattern_line
 
 
-def get_config_mapping(check, global_config):
-    attribute_path = config_mapping.get(check.check_subtype)
+def get_config_mapping(config_name, global_config):
+    attribute_path = config_mapping.get(config_name)
     if attribute_path:
         attrs = attribute_path.split(".")
         value = global_config
@@ -216,11 +216,13 @@ config_mapping = {
     "coredump": "config_files.coredump",
     "crypto-policies": "config_files.crypto_policies",
     "gpg": "config_files.gpg",
+    "kernel": "config_files.kernel",
     "selinux": "config_files.selinux",
     "shell": "config_files.shell",
     "sshd": "config_files.sshd",
     "sudo": "config_files.sudo",
     "sysctl": "config_files.sysctl",
+    "umask": "config_files.umask",
 }
 
 pkgmgr_mapping = {