+++ src/hardshell/common/logging.py
@@ -1,4 +1,3 @@
-# logging_config.py
 import logging
 
 # Configure logging
@@ -6,7 +5,10 @@ logging.basicConfig(
     level=logging.INFO,
     format="%(asctime)s - %(levelname)s - %(message)s",
     datefmt="%Y-%m-%d %H:%M:%S",
-    handlers=[logging.FileHandler("hardshell.log"), logging.StreamHandler()],
+    handlers=[
+        logging.FileHandler("hardshell.log"),
+        # logging.StreamHandler()
+    ],
 )
 
 # Get a logger instance