+++ src/hardshell/main.py
@@ -2,7 +2,6 @@
 # Imports
 #########################################################################################
 import click
-
 from src.hardshell import __version__
 from src.hardshell.scan import start_scanner