+++ src/hardshell/scan.py
@@ -4,7 +4,7 @@
 from src.hardshell.common.checks import create_checks
 from src.hardshell.common.common import detect_os
 from src.hardshell.common.config import GlobalConfig, load_config
-from src.hardshell.report import Report
+from src.hardshell.common.report import Report
 
 
 def start_scanner():
@@ -27,13 +27,16 @@ def start_scanner():
 
         # Run Checks
         for check in checks:
-            # print(f"Running Check: {check.check_name}")
+            print(f"Running Check: {check.check_name}")
             # print(check)
-            check.run_check()
-            # check.run_check(current_os=detected_os, global_config=global_config)
-            # if check.check_results:
-            #     print(f"Check: {check.check_name} - Result: {check.check_results}")
-            #     report.add_entry(result=check.check_results)
+            check.run_check(current_os=detected_os, global_config=global_config)
+            if check.check_results:
+                print(f"Check: {check.check_name} - Result: {check.check_results}")
+                for result in check.check_results:
+                    report.add_entry(result=result)
+
+        # Check Report
+        report.export_to_txt("report.txt")
 
     else:
         print("Unsupported OS")